claude-api-tools/requirements.txt
anthropic>=0.7.0
requests>=2.26.0
python-dotenv>=0.19.0
PyGithub>=2.1.1